requirements.txt
replicate==0.22.0
python-dotenv==1.0.0
streamlit==1.31.1 
goodfire==0.3.4
nest-asyncio==1.5.8